requirements.txt
pyfiglet
pyserial
colorama
keyboard
datetime
requests